pyproject.toml
[tool.poetry]
name = "ez-order"
version = "0.1.0"
description = "A WhatsApp chatbot system using FastAPI, Gemini-compatible client, and MongoDB"
authors = ["Your Name <usman678zafar@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
pymongo = "^4.6.1"
whatsapp-api-client-python = "^0.0.49"
whatsapp-chatbot-python = "0.9.6"
openai = "^1.43.0"
anthropic = "^0.30.0"
fastapi = "^0.110.0"
uvicorn = "^0.22.0"
pydantic = "^2.5.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"
email-validator = "^2.1.0"
openai-agents = "^0.1.0"
SpeechRecognition = "^3.10.0"
pydub = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
